Move Merkle tree API to separate component
