+++ core/lib/zksync_core/src/api_server/mod.rs
@@ -1,6 +1,8 @@
 // Everywhere in this module the word "block" actually means "miniblock".
+
 pub mod contract_verification;
 pub mod execution_sandbox;
 pub mod healthcheck;
+pub mod tree;
 pub mod tx_sender;
 pub mod web3;
