+++ core/lib/zksync_core/src/metadata_calculator/metrics.rs
@@ -2,7 +2,6 @@
 
 use vise::{
     Buckets, EncodeLabelSet, EncodeLabelValue, Family, Gauge, Histogram, LatencyObserver, Metrics,
-    Unit,
 };
 
 use std::time::{Duration, Instant};
@@ -173,22 +172,3 @@ impl MetadataCalculator {
         APP_METRICS.block_latency[&BlockStage::Tree].observe(Duration::from_secs(latency));
     }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EncodeLabelValue, EncodeLabelSet)]
-#[metrics(label = "method", rename_all = "snake_case")]
-pub(super) enum MerkleTreeApiMethod {
-    Info,
-    GetProofs,
-}
-
-/// Metrics for Merkle tree API.
-#[derive(Debug, Metrics)]
-#[metrics(prefix = "server_merkle_tree_api")]
-pub(super) struct MerkleTreeApiMetrics {
-    /// Server latency of the Merkle tree API methods.
-    #[metrics(buckets = Buckets::LATENCIES, unit = Unit::Seconds)]
-    pub latency: Family<MerkleTreeApiMethod, Histogram<Duration>>,
-}
-
-#[vise::register]
-pub(super) static API_METRICS: vise::Global<MerkleTreeApiMetrics> = vise::Global::new();
